translated to new OpenCL dialect * binary operations * distance matrix operations * axis transpose * Jaccard/Sorenen-Dice measurements + corresponding tests

+++ src/main/java/net/haesleinhuepf/clijx/advancedmath/exponential_2d_x.cl
@@ -10,7 +10,7 @@ __kernel void exponential_2d(
 
   const int2 pos = (int2){x,y};
 
-  const double input = (double)READ_src_IMAGE(src, sampler, pos).x;
+  const float input = (float)READ_src_IMAGE(src, sampler, pos).x;
 
   const IMAGE_dst_PIXEL_TYPE value = CONVERT_dst_PIXEL_TYPE(exp( input ));
